add increase/decrease functions
And other small improvements.

This feature request by @BCaesarleo expands the Motu AVB Faders to be able to increment up and down.  

You now have two options:

* Preset buttons with the key down and key up actions using an interval timer to increase/decrease
* Single-button presses that you repeatedly tap to change the level

OSC protocol currently doesn't support bi-directional communication, so, once a command is sent to set the fader to a known amount, we store that as a variable for subsequent manipulation.

Because of the one-way protocol, if the device is manipulated outside of Companion in any way, the module may not reflect the correct fader level.